Remove windows namd pipes dep
